+++ src/index.html
@@ -9,7 +9,7 @@
 <body>
     <div class="to-do-box">
         <div class="todays-to-do"><p>Today's To Do</p><i class="fa-solid fa-arrows-rotate"></i></div>
-        <form class="add-to-list" id="add-to-list">
+        <form class="add-to-list">
             <input type="text" name="task" id="task" placeholder="Add to list..." required>
             <button type="submit" class="btn-add-task">
                 <i class="fa-solid fa-plus"></i>
